+++ node/models/NuevosRepuestosModel.js
@@ -2,7 +2,7 @@ import db from "../db/RectimotorDB.js";
 import { DataTypes } from "sequelize";
 
 const NuevosRepuestosModel = db.define('NUEVOS_REPUESTOS', {
-    ID_REPUESTO  : { type: DataTypes.DECIMAL, primaryKey: true },
+    ID_REPUESTO  : { type: DataTypes.INTEGER, primaryKey: true },
     NOMBRE_REPUESTO   : { type: DataTypes.STRING}
 }, { tableName: 'NUEVOS_REPUESTOS' })
