+++ node/models/OrdenModel.js
@@ -2,9 +2,9 @@ import db from "../db/RectimotorDB.js";
 import { DataTypes } from "sequelize";
 
 const OrdenModel = db.define('ORDENES', {
-    id_orden: { type: DataTypes.DECIMAL, primaryKey: true },
-    id_motor: { type: DataTypes.DECIMAL },
-    id_taller: { type: DataTypes.DECIMAL },
+    id_orden: { type: DataTypes.INTEGER, primaryKey: true },
+    id_motor: { type: DataTypes.INTEGER },
+    id_taller: { type: DataTypes.INTEGER },
     cc_persona: { type: DataTypes.DECIMAL },
     placa: { type: DataTypes.STRING },
     fecha_recibido: { type: DataTypes.DATE },
