+++ node/models/PartesModel.js
@@ -2,7 +2,7 @@ import db from "../db/RectimotorDB.js";
 import { DataTypes } from "sequelize";
 
 const PartesModel = db.define('PARTES', {
-    ID_PARTE   : { type: DataTypes.DECIMAL, primaryKey: true },
+    ID_PARTE   : { type: DataTypes.INTEGER, primaryKey: true },
     NOMBRE_PARTE   : { type: DataTypes.STRING}
 }, { tableName: 'PARTES' })
